+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/AbstractSocialTokenVerifier.java
@@ -1,5 +1,6 @@
 package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
+import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
 import com.mobble.mobbleserver.global.exception.common.DomainException;
 import com.mobble.mobbleserver.global.exception.errorCode.oAuth2.OAuth2ErrorCode;
 import lombok.RequiredArgsConstructor;
