+++ src/main/java/com/mobble/mobbleserver/account/auth/oauth/verifier/SocialVerifier.java
@@ -1,5 +1,7 @@
 package com.mobble.mobbleserver.account.auth.oauth.verifier;
 
+import com.mobble.mobbleserver.account.auth.oauth.verifier.dto.SocialUserInfo;
+
 public interface SocialVerifier {
     SocialUserInfo verify(String accessToken);
 }
